register command properly
